refactor(backend): split web-ft.py into modular components
Split the monolithic web-ft.py (425 lines) into a modular Flask
application using the app factory pattern and Blueprints:

- app.py: Flask app factory with configuration and DB setup
- routes/views.py: Page-rendering routes (/, /map, /table, /display)
- routes/api.py: JSON API routes (/spots, /tbspots)
- services/spots.py: MongoDB query and data processing logic
- services/geo.py: CQ zone GeoJSON loading and point-in-polygon lookups
- utils/bands.py: Frequency-to-band conversion utility

web-ft.py is kept as a thin wrapper for backwards compatibility so
existing deployment commands (gunicorn web-ft:app, python web-ft.py)
continue to work unchanged.

Additional improvements:
- Deduplicated time-query logic into shared _build_time_query() helper
- Receiver grid now read from RECEIVER_GRID env var (prepares for #38)
- Added None guard to frequency_to_band() for missing frequency data
- Removed unused geopy and geopandas imports

Closes #33
