+++ utils/__init__.py
@@ -1,0 +1,1 @@
+"""Shared utility functions for the HamSCI Dashboard."""
